+++ src/bot.js
@@ -14,6 +14,11 @@ client.disabled = new Discord.Collection();
 client.cooldowns = new Discord.Collection();
 client.defaultPrefix = "w/";
 client.botVersion = "1.2.1";
+process.env.userAgent = "Discord Welcome-Bot " + client.botVersion;
+process.env.ownerIDs = [
+    "815204465937481749" /*PuneetGopinath#6398*/,
+    "693754859014324295" /*abhijoshi2k#6842*/,
+];
 
 const commandFolder = __dirname + "/commands";
 const commandFolders = fs.readdirSync(commandFolder);
@@ -22,6 +27,8 @@ let defaultOpts = {
     args: false,
     catchError: true,
     disabled: false,
+    cooldown: 3,
+    ownerOnly: false,
 };
 
 for (const folder of commandFolders) {
@@ -86,13 +93,13 @@ client.on("ready", () => {
 //https://discord.js.org/#/docs/main/v12/class/Client?scrollTo=e-guildMemberAdd
 client.on("guildMemberAdd", (member) => {
     // When a new member joins
-    greetUser(member.guild, member);
+    greetUser(member);
 });
 
 //https://discord.js.org/#/docs/main/v12/class/Client?scrollTo=e-guildMemberRemove
 client.on("guildMemberRemove", (member) => {
     // When a member leaves or is kicked or is banned
-    sayGoodBye(member.guild, member);
+    sayGoodBye(member);
 });
 
 //https://discord.js.org/#/docs/main/v12/class/Client?scrollTo=e-guildCreate