+++ src/commands/info/info.js
@@ -8,6 +8,7 @@ module.exports = {
     aliases: [],
     description: "Bot information",
     usage: "(--dm)",
+    cooldown: 10,
     execute(message, args) {
         const { MessageEmbed } = require("discord.js");
         let msg = new MessageEmbed();
@@ -17,14 +18,20 @@ module.exports = {
         msg.addField("Servers joined:", message.client.guilds.cache.size);
         msg.addField("Version:", `${message.client.botVersion}`);
         msg.addField(
+            "Invite URL:",
+            "[https://dsc.gg/welcome-bot2](https://dsc.gg/welcome-bot2)"
+        );
+        msg.addField(
             "Bot lists:",
-            `[discordextremelist.xyz](https://discordextremelist.xyz/en-US/bots/welcome-bot)`
+            `[discordextremelist.xyz](https://discordextremelist.xyz/en-US/bots/welcome-bot)\n` +
+                `[disbotlist.xyz](https://disbotlist.xyz/bot/848459799783669790)`
         );
         msg.addField(
             "Other links:",
             "[Support server](https://dsc.gg/welcome-bot-guild)\n" +
                 "[GitHub](https://github.com/Welcome-Bot/welcome-bot/)\n" +
-                "[Privacy policy](https://github.com/Welcome-Bot/welcome-bot/blob/main/docs/privacy-policy.md) and [Terms of service](https://github.com/Welcome-Bot/welcome-bot/blob/main/docs/terms.md)"
+                "[Privacy policy](https://welcome-bot.github.io/docs/privacy-policy.md) and [Terms of service](hhttps://welcome-bot.github.io/docs/terms.md)\n" +
+                "[Documentation](https://welcome-bot.github.io/docs)"
         );
         switch (args[0]) {
             case "--dm":