+++ src/commands/mod/ban.js
@@ -12,6 +12,7 @@ module.exports = {
     args: true,
     guildOnly: true,
     catchError: false,
+    cooldown: 5,
     usage: "[@user] (reason)",
     async execute(message, args) {
         const getUserFromMention = require("../../functions/getUserFromMention.js");