+++ src/functions/sayGoodBye.js
@@ -4,9 +4,9 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const getGuild = require("../db/functions/getGuild");
-module.exports = async (guild, member) => {
-    let guildDB = await getGuild(guild.id);
-    let channel = guild.channels.cache.find(
+module.exports = async (member) => {
+    let guildDB = await getGuild(member.guild.id);
+    let channel = member.guild.channels.cache.find(
         (ch) => ch.name === guildDB.welcomeChannel
     );
     if (!channel) {
@@ -17,8 +17,8 @@ module.exports = async (guild, member) => {
     //Replace Placeholders with their values
     msg = msg
         .replace("{mention}", `${member}`)
-        .replace("{server}", `${guild.name}`)
-        .replace("{members}", `${guild.memberCount}`);
+        .replace("{server}", `${member.guild.name}`)
+        .replace("{members}", `${member.guild.memberCount}`);
     channel.send(msg);
     channel.stopTyping();
 };